requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# NLP & ML dependencies
sentence-transformers==2.2.2
hdbscan==0.8.33
numpy==1.24.3
scikit-learn==1.3.2
torch>=2.0.0

# OpenAI API
openai==1.6.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1